+++ example/ios/Sharing Extension/ShareViewController.swift
@@ -21,6 +21,7 @@ class ShareViewController: SLComposeServiceViewController {
     let videoContentType = kUTTypeMovie as String
     let textContentType = kUTTypeText as String
     let urlContentType = kUTTypeURL as String
+    let fileURLType = kUTTypeFileURL as String;
     
     override func isContentValid() -> Bool {
         return true
@@ -28,15 +29,15 @@ class ShareViewController: SLComposeServiceViewController {
     
     override func viewDidLoad() {
         // This is called after the user selects Post. Do the upload of contentText and/or NSExtensionContext attachments.
-        
         if let content = extensionContext!.inputItems[0] as? NSExtensionItem {
             if let contents = content.attachments {
                 for (index, attachment) in (contents as! [NSItemProvider]).enumerated() {
-                    
                     if attachment.hasItemConformingToTypeIdentifier(imageContentType) {
                         handleImages(content: content, attachment: attachment, index: index)
                     } else if attachment.hasItemConformingToTypeIdentifier(textContentType) {
                         handleText(content: content, attachment: attachment, index: index)
+                    } else if attachment.hasItemConformingToTypeIdentifier(fileURLType) {
+                        handleFiles(content: content, attachment: attachment, index: index)
                     } else if attachment.hasItemConformingToTypeIdentifier(urlContentType) {
                         handleUrl(content: content, attachment: attachment, index: index)
                     } else if attachment.hasItemConformingToTypeIdentifier(videoContentType) {
@@ -161,10 +162,38 @@ class ShareViewController: SLComposeServiceViewController {
         }
     }
     
+    private func handleFiles (content: NSExtensionItem, attachment: NSItemProvider, index: Int) {
+        attachment.loadItem(forTypeIdentifier: fileURLType, options: nil) { [weak self] data, error in
+
+            if error == nil, let url = data as? URL, let this = self {
+
+                // Always copy
+                let newName = this.getFileName(from :url)
+                let newPath = FileManager.default
+                    .containerURL(forSecurityApplicationGroupIdentifier: "group.\(this.hostAppBundleIdentifier)")!
+                    .appendingPathComponent("\(newName)")
+                let copied = this.copyFile(at: url, to: newPath)
+                if (copied) {
+                    this.sharedMedia.append(SharedMediaFile(path: newPath.absoluteString, thumbnail: nil, duration: nil, type: .file))
+                }
+
+                if index == (content.attachments?.count)! - 1 {
+                    let userDefaults = UserDefaults(suiteName: "group.\(this.hostAppBundleIdentifier)")
+                    userDefaults?.set(this.toData(data: this.sharedMedia), forKey: this.sharedKey)
+                    userDefaults?.synchronize()
+                    this.redirectToHostApp(type: .file)
+                }
+
+            } else {
+                self?.dismissWithError()
+            }
+        }
+    }
+    
     private func dismissWithError() {
-        print("GETTING ERROR")
+        print("[ERROR] Error loading data!")
         let alert = UIAlertController(title: "Error", message: "Error loading data", preferredStyle: .alert)
-        
+
         let action = UIAlertAction(title: "Error", style: .cancel) { _ in
             self.dismiss(animated: true, completion: nil)
         }
@@ -175,7 +204,7 @@ class ShareViewController: SLComposeServiceViewController {
     }
     
     private func redirectToHostApp(type: RedirectType) {
-        let url = URL(string: "SharePhotos://dataUrl=\(sharedKey)#\(type)")
+        let url = URL(string: "ShareMedia://dataUrl=\(sharedKey)#\(type)")
         var responder = self as UIResponder?
         let selectorOpenURL = sel_registerName("openURL:")
         
@@ -191,6 +220,7 @@ class ShareViewController: SLComposeServiceViewController {
     enum RedirectType {
         case media
         case text
+        case file
     }
     
     func getExtension(from url: URL, type: SharedMediaType) -> String {
@@ -202,13 +232,25 @@ class ShareViewController: SLComposeServiceViewController {
         
         if (ex == nil) {
             switch type {
-            case .image:
-                ex = "PNG"
-            case .video:
-                ex = "MP4"
+                case .image:
+                    ex = "PNG"
+                case .video:
+                    ex = "MP4"
+                case .file:
+                    ex = "TXT"
             }
         }
         return ex ?? "Unknown"
+    }
+    
+    func getFileName(from url: URL) -> String {
+        var name = url.lastPathComponent
+        
+        if (name == "") {
+            name = UUID().uuidString + "." + getExtension(from: url, type: .file)
+        }
+        
+        return name
     }
     
     func copyFile(at srcURL: URL, to dstURL: URL) -> Bool {
@@ -218,7 +260,7 @@ class ShareViewController: SLComposeServiceViewController {
             }
             try FileManager.default.copyItem(at: srcURL, to: dstURL)
         } catch (let error) {
-            print("Cannot copy item at \(srcURL) to \(dstURL): \(error)")
+            print("Cannot copy item at \(srcURL) to \(dstURL): \(error)") 
             return false
         }
         return true
@@ -271,11 +313,17 @@ class ShareViewController: SLComposeServiceViewController {
             self.duration = duration
             self.type = type
         }
+
+        // Debug method to print out SharedMediaFile details in the console
+        func toString() {
+            print("[SharedMediaFile] \n\tpath: \(self.path)\n\tthumbnail: \(self.thumbnail)\n\tduration: \(self.duration)\n\ttype: \(self.type)")
+        }
     }
     
     enum SharedMediaType: Int, Codable {
         case image
         case video
+        case file
     }
     
     func toData(data: [SharedMediaFile]) -> Data {